+++ projects/ngx-material/src/lib/ngx-overlay/components/render-overlay/render-overlay.component.ts
@@ -1,11 +1,9 @@
-import {
-  ChangeDetectionStrategy, ChangeDetectorRef, Component, DoCheck, ElementRef, Inject, Injector, TemplateRef
-} from '@angular/core';
+import {ChangeDetectionStrategy, Component, ElementRef, Inject} from '@angular/core';
 import {Observable} from "rxjs";
-import {OVERLAY_ANIMATE_IN, OVERLAY_CONTEXT, OVERLAY_Z_INDEX} from "../../models/overlay-tokens.models";
-import {OverlayContext} from "../../models/overlay-context.models";
+import {OVERLAY_ANIMATE_IN} from "../../models/overlay-tokens.models";
 import {map} from "rxjs/operators";
-import {overlayAnimation} from '@consensus-labs/ngx-tools'
+import {overlayAnimation, TemplateRendering} from '@consensus-labs/ngx-tools'
+import {OverlayContext} from "../../models/overlay-context.models";
 
 @Component({
   templateUrl: './render-overlay.component.html',
@@ -16,41 +14,32 @@ import {overlayAnimation} from '@consensus-labs/ngx-tools'
   ],
   host: {'[@overlay]': 'animate'}
 })
-export class RenderOverlayComponent implements DoCheck {
+export class RenderOverlayComponent {
 
-  template: TemplateRef<void>;
-  onClose: () => any;
+  content: TemplateRendering;
   canClose$: Observable<boolean>;
   scrollable$: Observable<boolean>;
   maxWidth$: Observable<string>;
 
-  onChange?: () => any;
-
   constructor(
     element: ElementRef<HTMLElement>,
-    @Inject(OVERLAY_CONTEXT) context: OverlayContext,
-    @Inject(OVERLAY_Z_INDEX) zIndex: number|undefined,
-    @Inject(OVERLAY_ANIMATE_IN) public animate: boolean,
-    public injector: Injector,
-    private changes: ChangeDetectorRef
+    private context: OverlayContext,
+    @Inject(OVERLAY_ANIMATE_IN) public animate: boolean
   ) {
-    this.changes.detach();
-    element.nativeElement.style.zIndex = zIndex?.toFixed(0) ?? '';
 
-    this.template = context.template;
+    element.nativeElement.style.zIndex = context.zIndex?.toFixed(0) ?? '';
+
+    this.content = context.content;
+
     this.canClose$ = context.canClose$;
-    this.onClose = context.onClose;
     this.scrollable$ = context.scrollable$;
     this.maxWidth$ = context.maxWidth$.pipe(
       map(x => x ?? 1600),
       map(x => `${x}px`)
     );
-
-    this.onChange = context.onChange;
-    context.changes$?.subscribe(() => this.changes.detectChanges());
   }
 
-  ngDoCheck() {
-    this.onChange?.();
+  onClose() {
+    this.context.close();
   }
 }