+++ projects/ngx-material/src/lib/ngx-overlay/services/overlay-manager.service.ts
@@ -1,7 +1,7 @@
-import {Injectable} from '@angular/core';
+import {Injectable, TemplateRef, ViewContainerRef} from '@angular/core';
 import {Scheduler} from "@consensus-labs/rxjs-tools";
-import {OverlayContext} from "../models/overlay-context.models";
-import {OverlayService, OverlayToken} from "@consensus-labs/ngx-tools";
+import {OverlayInstance, OverlayOptions} from "../models/overlay-context.models";
+import {OverlayService} from "@consensus-labs/ngx-tools";
 
 @Injectable({providedIn: 'root'})
 export class OverlayManagerService {
@@ -13,21 +13,20 @@ export class OverlayManagerService {
   constructor(private overlayService: OverlayService) {
   }
 
-  createOverlay(overlay: OverlayContext): OverlayInstance {
+  createOverlay(
+    viewContainer: ViewContainerRef,
+    contentTemplate: TemplateRef<any>,
+    options: OverlayOptions
+  ): OverlayInstance {
     const token = this.overlayService.pushOverlay();
-    token.handleEscape(() => overlay.onClose?.());
-    const instance = {context: overlay, token};
+    const instance = new OverlayInstance(viewContainer, contentTemplate, token, options);
     this.scheduler.push(instance);
     return instance;
   }
 
   closeOverlay(instance: OverlayInstance) {
-    instance.token.dispose();
+    instance.dispose();
     this.scheduler.removeItem(instance);
   }
 }
 
-export interface OverlayInstance {
-  context: OverlayContext;
-  token: OverlayToken;
-}