+++ projects/ngx-tools/src/lib/rendering/index.ts
@@ -1,0 +1,4 @@
+
+export * from './models/template-rendering'
+export * from './directives/render-outlet.directive'
+export * from './ngx-rendering.module'