+++ projects/ngx-tools/src/lib/rendering/ngx-rendering.module.ts
@@ -1,0 +1,9 @@
+import {NgModule} from '@angular/core';
+import {RenderOutletDirective} from "./directives/render-outlet.directive";
+
+@NgModule({
+  exports: [RenderOutletDirective],
+  declarations: [RenderOutletDirective],
+})
+export class NgxRenderingModule {
+}