+++ projects/ngx-tools/src/public-api.ts
@@ -4,6 +4,7 @@
 
 export * from './lib/future-builder';
 export * from './lib/icon';
+export * from './lib/rendering';
 
 export * from './services';
 export * from './directives';